requirements.txt
# Core ML / DL
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
accelerate>=0.24.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# NLP
tokenizers>=0.15.0
sentencepiece>=0.1.99

# Explainability
shap>=0.43.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars & utilities
tqdm>=4.65.0

# Jupyter
jupyter>=1.0.0
jupyterlab>=4.0.0
ipykernel>=6.25.0
ipywidgets>=8.0.0
nbformat>=5.9.0

# Data download
requests>=2.31.0
gdown>=4.7.0

# Configuration
pyyaml>=6.0.1
omegaconf>=2.3.0

# Logging
tensorboard>=2.14.0

# Statistics
statsmodels>=0.14.0
pingouin>=0.5.3

# Type hints
typing-extensions>=4.8.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
